requirements-dev.txt
-r requirements.txt
django-debug-toolbar
factory-boy
ipdb
mypy
prospector
pytest
pytest-django
pytest-pythonpath
